+++ bootique-metrics-web/src/main/java/io/bootique/metrics/web/MetricsWebModuleProvider.java
@@ -1,0 +1,15 @@
+package io.bootique.metrics.web;
+
+import com.google.inject.Module;
+import io.bootique.BQModuleProvider;
+
+/**
+ * @since 0.8
+ */
+public class MetricsWebModuleProvider implements BQModuleProvider {
+
+    @Override
+    public Module module() {
+        return new MetricsWebModule();
+    }
+}