+++ bootique-metrics/src/main/java/io/bootique/metrics/MetricsModule.java
@@ -1,7 +1,6 @@
 package io.bootique.metrics;
 
 import com.codahale.metrics.MetricRegistry;
-import com.codahale.metrics.health.HealthCheck;
 import com.google.inject.Binder;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
@@ -9,6 +8,7 @@ import com.google.inject.multibindings.MapBinder;
 import com.google.inject.multibindings.Multibinder;
 import io.bootique.ConfigModule;
 import io.bootique.config.ConfigurationFactory;
+import io.bootique.metrics.healthcheck.HealthCheck;
 import io.bootique.metrics.healthcheck.HealthCheckGroup;
 import io.bootique.metrics.healthcheck.HealthCheckRegistry;